// lib/api/generated/models/MemberResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ClassHub API
 * ClassHub Backend API 문서
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 회원 응답
 * @export
 * @interface MemberResponseDto
 */
export interface MemberResponseDto {
    /**
     * 회원 ID
     * @type {number}
     * @memberof MemberResponseDto
     */
    id?: number;
    /**
     * 회원 이름
     * @type {string}
     * @memberof MemberResponseDto
     */
    name?: string;
    /**
     * 연락처
     * @type {string}
     * @memberof MemberResponseDto
     */
    phone?: string;
    /**
     * 생성 일시
     * @type {Date}
     * @memberof MemberResponseDto
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the MemberResponseDto interface.
 */
export function instanceOfMemberResponseDto(value: object): value is MemberResponseDto {
    return true;
}

export function MemberResponseDtoFromJSON(json: any): MemberResponseDto {
    return MemberResponseDtoFromJSONTyped(json, false);
}

export function MemberResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MemberResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'phone': json['phone'] == null ? undefined : json['phone'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function MemberResponseDtoToJSON(json: any): MemberResponseDto {
    return MemberResponseDtoToJSONTyped(json, false);
}

export function MemberResponseDtoToJSONTyped(value?: MemberResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'phone': value['phone'],
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
    };
}